SurfaceFluxes.jl

SurfaceFluxes — Module

SurfaceFluxes

Interface

surface_conditions computes

Monin-Obukhov length

Potential temperature flux (if not given) using Monin-Obukhov theory

transport fluxes using Monin-Obukhov theory

friction velocity/temperature scale/tracer scales

exchange coefficients

References

S Nishizawa, Y Kitamura (2018)

Daewon W. Byun (1990)

Core input types

SurfaceFluxes.SurfaceValues — Type

SurfaceValues

Input container for state variables at the ground level.

Fields

z

u

ts

SurfaceFluxes.InteriorValues — Type

InteriorValues

Input container for state variables at the first interior node.

Fields

z

u

ts

Dispatch types

SurfaceFluxes.Fluxes — Type

Fluxes

Input container for state variables, latent and sensible heat fluxes roughness lengths, initial obukhov length and gustiness.

Fields

state_in

state_sfc

shf

lhf

z0m

z0b

L_MO_init

gustiness

SurfaceFluxes.FluxesAndFrictionVelocity — Type

FluxesAndFrictionVelocity

Input container, given surface state variables, latent and sensible heat fluxes, and the friction velocity, roughness lengths, initial obukhov length and gustiness.

Fields

state_in

state_sfc

shf

lhf

ustar

z0m

z0b

gustiness

SurfaceFluxes.Coefficients — Type

Coefficients

Input container, given surface state variables, and exchange coefficients,roughness lengths, initial obukhov length and gustiness.

Fields

state_in

state_sfc

Cd

Ch

z0m

z0b

gustiness

SurfaceFluxes.ValuesOnly — Type

ValuesOnly

Input container, given only surface state variables, roughness lengths, initial obukhov length and gustiness.

Fields

state_in

state_sfc

z0m

z0b

L_MO_init

gustiness

User-facing methods

SurfaceFluxes.surface_conditions — Function

surface_conditions(
    param_set::CLIMAParameters.AbstractParameterSet,
    sc::SurfaceFluxes.AbstractSurfaceConditions{FT},
    scheme::SurfaceFluxes.SolverScheme = FVScheme();
    tol::RootSolvers.AbstractTolerance =
        RootSolvers.SolutionTolerance(max(FT(1e-1), sqrt(eps(FT)))),
    maxiter::Int = 10,
) where {FT}

The main user facing function of the module. It computes the surface conditions based on the Monin-Obukhov similarity functions. Requires information about thermodynamic parameters (param_set) the surface state sc, the universal function type and the discretisation scheme. Default tolerance for Monin-Obukhov length is absolute (i.e. has units [m]).

Result struct of type SurfaceFluxConditions{FT} contains:

L_MO: Monin-Obukhov lengthscale

shf: Sensible Heat Flux

lhf: Latent Heat Flux

ρτxz: Momentum Flux (Eastward component)

ρτyz: Momentum Flux (Northward component)

ustar: Friction velocity

Cd: Momentum Exchange Coefficient

Ch: Thermal Exchange Coefficient

SurfaceFluxes.recover_profile — Function

recover_profile(param_set, sc, L_MO, Z, X_in, X_sfc, transport, uft, scheme)

Recover profiles of variable X given values of Z coordinates. Follows Nishizawa equation (21,22)

Arguments

param_set: Abstract Parameter Set containing physical, thermodynamic parameters.

sc: Container for surface conditions based on known combination of the state vector, and {fluxes, friction velocity, exchange coefficients} for a given experiment

L_MO: Monin-Obukhov length

Z: Z coordinate(s) (within surface layer) for which variable values are required

Xin,Xsfc: For variable X, values at interior and surface nodes

transport: Transport type, (e.g. Momentum or Heat, used to determine physical scale coefficients)

uft: A Universal Function type, (returned by, e.g., Businger())

scheme: Discretization scheme (currently supports FD and FV)

TODO: add tests

Universal Functions

SurfaceFluxes.UniversalFunctions — Module

UniversalFunctions

Universal stability and stability correction functions for SurfaceFluxes module. Supports universal functions:

Businger

Gryanik

Grachev

SurfaceFluxes.UniversalFunctions.Gryanik — Type

Gryanik <: AbstractUniversalFunction{FT}

References

Vladimir M Gryanik, Christof L{\"u}pkes, Andrey Grachev, Dmitry Sidorenko (2020)

Equations in reference:

`ϕ_m`: Eq. 13
`ϕ_h`: Eq. 13
`ψ_m`: Eq. 14
`ψ_h`: Eq. 14

Fields

L

Monin-Obhukov Length

params

SurfaceFluxes.UniversalFunctions.Grachev — Type

Grachev <: AbstractUniversalFunction{FT}

References

Andrey A Grachev, Edgar L Andreas, Christopher W Fairall, Peter S Guest, P Ola G Persson (2007)

Equations in reference:

`ϕ_m`: Eq. 13
`ϕ_h`: Eq. 13
`ψ_m`: Eq. 14
`ψ_h`: Eq. 14

Fields

L

Monin-Obhukov Length

params

SurfaceFluxes.UniversalFunctions.Businger — Type

Businger

Reference

S Nishizawa, Y Kitamura (2018)

Original research

Joost A Businger, John C Wyngaard, Yꎬ Izumi, Edward F Bradley (1971)

Equations in reference:

`ϕ_m`: Eq. A1
`ϕ_h`: Eq. A2
`ψ_m`: Eq. A3
`ψ_h`: Eq. A4

Fields

L

Monin-Obhukov Length

params

SurfaceFluxes.UniversalFunctions.phi — Function

phi

Universal stability function for wind shear (ϕ_m) and temperature gradient (ϕ_h)

SurfaceFluxes.UniversalFunctions.psi — Function

psi

Universal stability correction function for momentum (ψ_m) and heat (ψ_h)

SurfaceFluxes.UniversalFunctions.Psi — Function

Psi

Integral of universal stability correction function for momentum (ψ_m) and heat (ψ_h)