SurfaceFluxes.jl

SurfaceFluxes — Module

SurfaceFluxes

Interface

surface_conditions computes

Monin-Obukhov length

Potential temperature flux (if not given) using Monin-Obukhov theory

transport fluxes using Monin-Obukhov theory

friction velocity/temperature scale/tracer scales

exchange coefficients

References

S Nishizawa, Y Kitamura (2018)

Universal Functions

SurfaceFluxes.UniversalFunctions — Module

UniversalFunctions

Universal stability and stability correction functions for SurfaceFluxes module.

SurfaceFluxes.UniversalFunctions.Gryanik — Type

Gryanik <: AbstractUniversalFunction{FT}

References

Vladimir M Gryanik, Christof L{\"u}pkes, Andrey Grachev, Dmitry Sidorenko (2020)

Equations in reference:

ϕ_m: Eq. 13
ϕ_h: Eq. 13
ψ_m: Eq. 14
ψ_h: Eq. 14

Fields

param_set

Parameter set

L

Monin-Obhukov Length

SurfaceFluxes.UniversalFunctions.Grachev — Type

Grachev <: AbstractUniversalFunction{FT}

References

Andrey A Grachev, Edgar L Andreas, Christopher W Fairall, Peter S Guest, P Ola G Persson (2007)

Equations in reference:

ϕ_m: Eq. 13
ϕ_h: Eq. 13
ψ_m: Eq. 14
ψ_h: Eq. 14

Fields

param_set

Parameter set

L

Monin-Obhukov Length

SurfaceFluxes.UniversalFunctions.Businger — Type

Businger

Reference

S Nishizawa, Y Kitamura (2018)

Original research

Joost A Businger, John C Wyngaard, Yꎬ Izumi, Edward F Bradley (1971)

Equations in reference:

ϕ_m: Eq. A1
ϕ_h: Eq. A2
ψ_m: Eq. A3
ψ_h: Eq. A4

Fields

param_set

Parameter set

L

Monin-Obhukov Length

SurfaceFluxes.UniversalFunctions.phi — Function

phi

Universal stability function for wind shear (ϕm) and temperature gradient (ϕh)

SurfaceFluxes.UniversalFunctions.psi — Function

psi

Universal stability correction function for momentum (ψm) and heat (ψh)

SurfaceFluxes.UniversalFunctions.Psi — Function

Psi

Integral of universal stability correction function for momentum (ψm) and heat (ψh)

Fluxes API

SurfaceFluxes.surface_conditions — Function

surface_conditions

Surface conditions given

MO_param_guess initial guess for solution (L_MO, u_star, θ_star, ϕ_star, ...)

x_in Inner values for state variable array x. They correspond to volume averages in FV and nodal point values in DG

x_s surface values for state variable array x

z_0 roughness lengths for state variable array x

θ_scale virtual dry static energy (i.e., basic potential temperature)

z_in Input height for the similarity functions. It is Δz for FV and the height of the first nodal point for DG

wθ_flux_star potential temperature flux (optional)

universal_func family of flux-gradient universal functions used (optional)

If wθ_flux_star is not given, then it is computed by iteration of equations 3, 17, and 18 in Nishizawa2018.

SurfaceFluxes.exchange_coefficients — Function

exchange_coefficients(
    param_set,
    z,
    F_exchange,
    x_star::VFT,
    θ_scale,
    L_MO,
    universal_func,
)

Computes exchange transfer coefficients

K_exchange exchange coefficients

SurfaceFluxes.recover_profile — Function

recover_profile(z, x_star, x_s, z_0, L_MO, transport, ::DGScheme, uf)

Recover vertical profiles u(z), θ(z), ... using equations (4) and (5) for DG and (12), (13) for FV from Nishizawa & Kitamura (2018).

Arguments

z Input height for evaluation

x_star Surface layer scales for state variable array x.

x_s surface values for state variable array x

z_0 roughness lengths for state variable array x

L_MO Obukhov length

transport Type of transport (momentum or heat)

uf universal function family

SurfaceFluxes.monin_obukhov_length — Function

monin_obukhov_length(param_set, u_star, θ_scale, wθ_surf_flux)

Returns the Monin-Obukhov length.