SurfaceFluxes.jl

SurfaceFluxes — Module

SurfaceFluxes

Interface

surface_conditions computes

Monin-Obukhov length

Potential temperature flux (if not given) using Monin-Obukhov theory

transport fluxes using Monin-Obukhov theory

friction velocity/temperature scale/tracer scales

exchange coefficients

References

[1]

[2]

Core input types

SurfaceFluxes.StateValues — Type

StateValues

Input container for state variables at either first / interior nodes.

Fields

z

u

ts

Dispatch types

SurfaceFluxes.Fluxes — Type

Fluxes

Input container for state variables, latent and sensible heat fluxes roughness lengths, initial obukhov length and gustiness.

Fields

state_in

state_sfc

shf

lhf

z0m

z0b

gustiness

SurfaceFluxes.FluxesAndFrictionVelocity — Type

FluxesAndFrictionVelocity

Input container, given surface state variables, latent and sensible heat fluxes, and the friction velocity, roughness lengths, initial obukhov length and gustiness.

Fields

state_in

state_sfc

shf

lhf

ustar

z0m

z0b

gustiness

SurfaceFluxes.Coefficients — Type

Coefficients

Input container, given surface state variables, and exchange coefficients,roughness lengths, initial obukhov length and gustiness.

Fields

state_in

state_sfc

Cd

Ch

gustiness

beta

SurfaceFluxes.ValuesOnly — Type

ValuesOnly

Input container, given only surface state variables, roughness lengths, initial obukhov length and gustiness.

Fields

state_in

state_sfc

z0m

z0b

gustiness

beta

User-facing methods

SurfaceFluxes.surface_conditions — Function

surface_conditions(
    param_set::AbstractSurfaceFluxesParameters,
    sc::SurfaceFluxes.AbstractSurfaceConditions{FT},
    scheme::SurfaceFluxes.SolverScheme = PointValueScheme();
    tol::RS.AbstractTolerance = RS.RelativeSolutionTolerance(FT(0.01)),
    tol_neutral::FT = SFP.cp_d(param_set) / 100,
    maxiter::Int = 10,
    soltype::RS.SolutionType = RS.CompactSolution(),
    noniterative_stable_sol::Bool=true,
) where {FT}

The main user facing function of the module. It computes the surface conditions based on the Monin-Obukhov similarity functions. Requires information about thermodynamic parameters (param_set) the surface state sc, the universal function type and the discretisation scheme. Default tolerance for Monin-Obukhov length is absolute (i.e. has units [m]). Returns the RootSolvers CompactSolution by default.

Result struct of type SurfaceFluxConditions{FT} contains:

L_MO: Monin-Obukhov lengthscale

shf: Sensible Heat Flux

lhf: Latent Heat Flux

ρτxz: Momentum Flux (Eastward component)

ρτyz: Momentum Flux (Northward component)

ustar: Friction velocity

Cd: Momentum Exchange Coefficient

Ch: Thermal Exchange Coefficient

SurfaceFluxes.recover_profile — Function

recover_profile(param_set, sc, L_MO, Z, X_in, X_sfc, transport, uft, scheme)

Recover profiles of variable X given values of Z coordinates. Follows Nishizawa equation (21,22)

Arguments

param_set: Abstract Parameter Set containing physical, thermodynamic parameters.

sc: Container for surface conditions based on known combination of the state vector, and {fluxes, friction velocity, exchange coefficients} for a given experiment

L_MO: Monin-Obukhov length

Z: Z coordinate(s) (within surface layer) for which variable values are required

X_star: Scale parameter for variable X

X_sfc: For variable X, values at surface nodes

transport: Transport type, (e.g. Momentum or Heat, used to determine physical scale coefficients)

uft: A Universal Function type, (returned by, e.g., Businger())

scheme: Discretization scheme (currently supports FD and FV)

TODO: add tests

Universal Functions

SurfaceFluxes.UniversalFunctions — Module

UniversalFunctions

Universal stability and stability correction functions for SurfaceFluxes module. Supports universal functions:

Businger

Gryanik

Grachev

Beljaars

'Cheng'

Holtslag

SurfaceFluxes.UniversalFunctions.Gryanik — Type

Gryanik <: AbstractUniversalFunction{FT}

References

[3]

Equations in reference:

`ϕ_m`: Eq. 32
`ϕ_h`: Eq. 33
`ψ_m`: Eq. 34
`ψ_h`: Eq. 35

Gryanik et al. (2020) functions are used in stable conditions

In unstable conditions the functions of Businger (1971) are

assigned by default.

Fields

L: Monin-Obhukov Length

params

SurfaceFluxes.UniversalFunctions.Grachev — Type

Grachev <: AbstractUniversalFunction{FT}

References

[4]

Equations in reference:

`ϕ_m`: Eq. 12
`ϕ_h`: Eq. 12
`ψ_m`: Eq. 13
`ψ_h`: Eq. 13

Grachev (2007) functions are applicable in the

stable b.l. regime (ζ >= 0). Businger (1971) functions

are applied in the unstable b.l. (ζ<0) regime by

default.

Fields

L: Monin-Obhukov Length

params

SurfaceFluxes.UniversalFunctions.Businger — Type

Businger

Reference

[1]

Original research

[5]

Equations in reference:

`ϕ_m`: Eq. A1
`ϕ_h`: Eq. A2
`ψ_m`: Eq. A3
`ψ_h`: Eq. A4

Fields

L: Monin-Obhukov Length

params

SurfaceFluxes.UniversalFunctions.Beljaars — Type

Beljaars <: AbstractUniversalFunction{FT}

References

[6]

Equations in reference:

`ϕ_m`: Derived from Eq. 28
`ϕ_h`: Derived from Eq. 32
`ψ_m`: Eq. 28
`ψ_h`: Eq. 32

Beljaars and Holtslag (1991) functions are used in stable conditions

In unstable conditions the functions of Businger (1971) are

assigned by default.

Fields

L: Monin-Obhukov Length

params

SurfaceFluxes.UniversalFunctions.Cheng — Type

Cheng <: AbstractUniversalFunction{FT}

References

[7]

Equations in reference:

`ϕ_m`: Eq. 22
`ϕ_h`: Eq. 24
`ψ_m`: Eq. 21
`ψ_h`: Eq. 23

Cheng qnd Brutsaert (2005) functions are applicable in the

stable b.l. regime (ζ >= 0). Businger (1971) functions

are applied in the unstable b.l. (ζ<0) regime by

default.

Fields

L: Monin-Obhukov Length

params

SurfaceFluxes.UniversalFunctions.Holtslag — Type

Holtslag <: AbstractUniversalFunction{FT}

References

[8]

Equations in reference:

`ϕ_m`: Derived from Eq. 12
`ϕ_h`: Derived from Eq. 12
`ψ_m`: Eq. 12
`ψ_h`: Eq. 12

Holtslag and Bruin 1988 functions are used in stable conditions

In unstable conditions the functions of Businger (1971) are

assigned by default.

Fields

L: Monin-Obhukov Length

params

SurfaceFluxes.UniversalFunctions.phi — Function

phi

Universal stability function for wind shear (ϕ_m) and temperature gradient (ϕ_h)

SurfaceFluxes.UniversalFunctions.psi — Function

psi

Universal stability correction function for momentum (ψ_m) and heat (ψ_h)

SurfaceFluxes.UniversalFunctions.Psi — Function

Psi

Integral of universal stability correction function for momentum (ψ_m) and heat (ψ_h)